feat: extract session types + message-v2 to @opencode-ai/state
- Moved ProviderError from opencode to @opencode-ai/core/provider/error
- Moved NotFoundError + Provider.Model to @opencode-ai/core
- Moved isMedia to @opencode-ai/core/util/media
- Moved message-v2.ts to @opencode-ai/state (all @/ deps resolved)
- Updated 6 files to import from @opencode-ai/core instead of @/
- Created manual node_modules symlink for workspace resolution
- No typecheck regressions
