+++ packages/core/package.json
@@ -60,6 +60,7 @@
     "drizzle-kit": "catalog:"
   },
   "dependencies": {
+    "ai": "catalog:",
     "@ai-sdk/alibaba": "1.0.17",
     "@ai-sdk/amazon-bedrock": "4.0.112",
     "@ai-sdk/anthropic": "3.0.82",
