+++ packages/opencode/package.json
@@ -89,6 +89,7 @@
     "@opencode-ai/script": "workspace:*",
     "@opencode-ai/sdk": "workspace:*",
     "@opencode-ai/server": "workspace:*",
+    "@opencode-ai/state": "workspace:*",
     "@opencode-ai/tui": "workspace:*",
     "@openrouter/ai-sdk-provider": "2.9.0",
     "@opentelemetry/api": "1.9.0",
