+++ packages/opencode/test/plugin/openai-ws.test.ts
@@ -4,7 +4,7 @@ import { createServer, type IncomingMessage, type Server as HttpServer } from "n
 import net, { type AddressInfo, type Socket } from "node:net"
 import WebSocket, { WebSocketServer } from "ws"
 import { APICallError } from "ai"
-import { ProviderError } from "../../src/provider/error"
+import { ProviderError } from "@opencode-ai/core/provider/error"
 import { OpenAIWebSocket } from "../../src/plugin/openai/ws"
 import { OpenAIWebSocketPool, TITLE_HEADER } from "../../src/plugin/openai/ws-pool"
