+++ packages/opencode/test/session/retry.test.ts
@@ -7,7 +7,7 @@ import { Effect, Layer, Schedule, Schema } from "effect"
 import { CrossSpawnSpawner } from "@opencode-ai/core/cross-spawn-spawner"
 import { SessionRetry } from "../../src/session/retry"
 import { MessageV2 } from "../../src/session/message-v2"
-import { ProviderError } from "../../src/provider/error"
+import { ProviderError } from "@opencode-ai/core/provider/error"
 import { SessionID } from "../../src/session/schema"
 import { SessionStatus } from "../../src/session/status"
 import { testEffect } from "../lib/effect"
